Update dependencies in package-lock.json and fix minor error in App.jsx
Updated dependencies in the 'package-lock.json' and 'yarn.lock' files were due to the npm updating the packages, which ensures the latest security fixes are included. A minor syntax error was also fixed in App.jsx file where the semi-colon was replaced with the closing tag. This is to ensure the code aligns with the syntax rules of JSX.